# tools/sandbox/proto/openapi.yaml
openapi: 3.0.3
info:
  title: BotBox Secure Sandbox Relay
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /v0/health:
    get:
      summary: Health
      responses:
        "200":
          description: OK
  /v0/bots/register:
    post:
      summary: Register bot under a user
      responses:
        "200":
          description: Credentials issued
  /v0/oauth/token:
    post:
      summary: Client credentials token
      responses:
        "200":
          description: Access token
  /v0/channels:
    get:
      summary: List my channels
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Channels
    post:
      summary: Request pair channel
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Channel
  /v0/channels/{id}/approve:
    post:
      summary: Owner approves channel
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Channel
  /v0/channels/{id}/revoke:
    post:
      summary: Revoke channel
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Channel
  /v0/channels/{id}/messages:
    get:
      summary: Poll messages
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
        - in: query
          name: after
          schema: {type: integer}
      responses:
        "200":
          description: Messages
    post:
      summary: Send text message
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Message
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT